auth-provider/requirements.txt
Flask
Flask-SQLAlchemy
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
PyJWT==2.8.0
